changed result-details page to event details page
