+++ src/components/result-page/index.js
@@ -1,10 +1,10 @@
-import MapPage from '../map-page';
-import ResultDetails from './result-details';
+// import MapPage from '../map-page';
+// import ResultDetails from './result-details';
 
-function ResultPage() {
-  return (
-    <MapPage Component={ResultDetails} />
-  );
-}
+// function ResultPage() {
+//   return (
+//     <MapPage Component={ResultDetails} />
+//   );
+// }
 
-export default ResultPage
+// export default ResultPage
